Add paging support for realmode memory access
The first 1MiB of physical memory (also called realmode memory)
contains regions which are mapped to I/O and BIOS. It may be
desirable to access some of this memory, e.g., when writing to
video memory.
To allow this kind of access, the 1 MiB linear memory region starting
at 0xFFFF800000000000 has been paged to the realmode memory.