Fix size constraints on physical mmap
The worst case size of the physical memory map could be 0x8000000. This would
occur if 64GiB of physical memory were mapped in 1-page regions.
As the real size is likely much smaller, a region of kernel memory was reserved
for the physical memory map, to facilitate expansion.

To further reduce the required space, the responsibility of keeping track of
allocated physical memory was shifted to the calling code. This removes the
need for a seperate memory map to keep track of already allocated physical
memory.